feat: add game module
