+++ client/src/main.ts
@@ -1,5 +1,3 @@
-import { MapSize, deserializeTypedArrays } from './mesh';
-import { drawCells } from './drawCells';
 import { createUI } from './ui';
 import {
   loadOrGetMesh,
@@ -8,21 +6,21 @@ import {
   currentCellBiomes,
   currentCellCount,
   showError,
-  meshData,
-  biomeConfig,
-  setCurrentCellBiomes,
-  setCurrentCellCount,
-  setCurrentMapSize,
 } from './terrain';
 import { WIDTH, HEIGHT, SERVER_BASE_URL } from './config';
-import {
-  initializeWebSocket,
-  addToRoom,
-  removeFromRoom,
-  sendGameAction,
-  closeWebSocket,
-} from './network';
+import { initializeWebSocket, closeWebSocket } from './network';
 import { showGameNotification } from './notifications';
+import {
+  initGame,
+  joinGameRoom,
+  leaveGameRoom,
+  processGameData,
+  handleActionResult,
+  handleGameUpdate,
+  handleGameError,
+  currentGameId,
+  currentPlayerName,
+} from './game';
 
 const canvas = document.createElement('canvas');
 const container = document.getElementById('canvas-container')!;
@@ -39,7 +37,8 @@ async function initializeApp() {
   // Create UI first
   createUI(ctx);
 
-  setupGameplayEventListeners();
+  // Initialize gameplay handlers
+  initGame(canvas, ctx);
 
   // Start preloading all meshes in background (non-blocking)
   preloadMeshes();
@@ -468,51 +467,6 @@ function updatePlayersList(players: string[], currentPlayerName?: string) {
 }
 
 // Initialize WebSocket connection
-
-// Add function to handle action results
-function handleActionResult(data: any): void {
-  console.log('Action result received:', data);
-  
-  if (data.success) {
-    // showGameNotification(data.message || 'Action completed successfully', 'success');
-  } else {
-    showGameNotification(data.error || 'Action failed', 'error');
-  }
-}
-
-function handleGameUpdate(data: any): void {
-  console.log('Game update received:', data);
-  
-  // Update current game state
-  if (data.gameId === currentGameId) {
-    const gameState = data.state;
-    
-    // Check if turn changed
-    const turnChanged = gameState.currentPlayer !== undefined && 
-                       (gameState.currentPlayer !== (isMyTurn ? currentPlayerName : ''));
-    
-    // Update turn information
-    if (gameState.currentPlayer !== undefined) {
-      isMyTurn = gameState.currentPlayer === currentPlayerName;
-      updateTurnIndicator(gameState.currentPlayer, gameState.turnNumber);
-    }
-    
-    // Update territory data if provided
-    if (gameState.cellOwnership) {
-      currentTerritoryData = gameState.cellOwnership;
-    }
-    
-    // Update entity data if provided
-    if (gameState.entities) {
-      currentGameEntities = gameState.entities;
-      console.log('Updated entities:', currentGameEntities);
-    }
-    
-    // Re-render the game state to show new entities/changes
-    renderGameState();
-  }
-}
-
 // Handle player joining
 function handlePlayerJoined(data: { gameId: string, players: string[], newPlayer: string }) {
   console.log('Player joined:', data);
@@ -544,12 +498,6 @@ function handleGameStarted(gameData: any) {
     updateGameStatus('in_progress');
     showGameNotification('Game has started!', 'success');
 
-    // Check if it's my turn
-    isMyTurn = gameData.state.currentPlayer === currentPlayerName;
-    
-    // Update turn indicator
-    updateTurnIndicator(gameData.state.currentPlayer, gameData.state.turnNumber);
-
     const startButton = document.getElementById("startGame") as HTMLButtonElement;
     if (startButton) startButton.remove();
 
@@ -559,327 +507,6 @@ function handleGameStarted(gameData: any) {
     // Process all the game data received in the WebSocket event
     processGameData(gameData);
   }
-}
-
-function updateTurnIndicator(currentPlayer: string, turnNumber: number): void {
-  let turnIndicator = document.getElementById('turnIndicator');
-  
-  if (!turnIndicator) {
-    // Create turn indicator element
-    turnIndicator = document.createElement('div');
-    turnIndicator.id = 'turnIndicator';
-    turnIndicator.style.cssText = `
-      position: fixed;
-      top: 70px;
-      left: 50%;
-      transform: translateX(-50%);
-      padding: 15px 25px;
-      border-radius: 8px;
-      color: white;
-      font-family: Arial, sans-serif;
-      font-size: 16px;
-      font-weight: bold;
-      z-index: 1500;
-      text-align: center;
-      min-width: 250px;
-      display: flex;
-      flex-direction: column;
-      gap: 10px;
-    `;
-    document.body.appendChild(turnIndicator);
-  }
-  
-  if (isMyTurn) {
-    turnIndicator.style.background = 'rgba(76, 175, 80, 0.9)';
-    turnIndicator.innerHTML = `
-      <div>Your Turn - Turn ${turnNumber}</div>
-      <button id="endTurnButton" style="
-        background: #4CAF50;
-        color: white;
-        border: solid 3px rgba(80, 182, 80, 0.9);;
-        padding: 8px 16px;
-        border-radius: 4px;
-        cursor: pointer;
-        font-size: 14px;
-        font-weight: bold;
-        transition: background-color 0.2s;
-      ">End Turn</button>
-    `;
-    
-    // Add event listener to the end turn button
-    const endTurnButton = document.getElementById('endTurnButton');
-    if (endTurnButton) {
-      endTurnButton.addEventListener('click', endTurn);
-      endTurnButton.addEventListener('mouseenter', (e) => {
-        (e.target as HTMLElement).style.background = '#45a049';
-      });
-      endTurnButton.addEventListener('mouseleave', (e) => {
-        (e.target as HTMLElement).style.background = '#4CAF50';
-      });
-    }
-  } else {
-    turnIndicator.style.background = 'rgba(255, 193, 7, 0.9)';
-    turnIndicator.innerHTML = `
-      <div>Waiting for ${currentPlayer} - Turn ${turnNumber}</div>
-    `;
-  }
-}
-
-function endTurn(): void {
-  if (!isMyTurn) {
-    showGameNotification('It is not your turn', 'error');
-    return;
-  }
-
-  const endTurnButton = document.getElementById('endTurnButton') as HTMLButtonElement;
-  if (endTurnButton) {
-    endTurnButton.disabled = true;
-    endTurnButton.textContent = 'Ending Turn...';
-    endTurnButton.style.background = '#cccccc';
-  }
-
-  deselectUnit();
-
-  dispatchGameAction('end_turn', {});
-
-  console.log(`${currentPlayerName} ending turn`);
-}
-
-function handleCellClick(event: MouseEvent): void {
-  if (!isMyTurn || !meshData || currentGameTerrain?.length === 0) {
-    return;
-  }
-
-  const rect = canvas.getBoundingClientRect();
-  const x = event.clientX - rect.left;
-  const y = event.clientY - rect.top;
-
-  // Find which cell was clicked
-  const clickedCellId = findCellAtPosition(x, y);
-  
-  if (clickedCellId !== -1) {
-    // Check if there's a unit on this cell
-    const unitOnCell = findUnitOnCell(clickedCellId);
-    
-    if (unitOnCell && unitOnCell.owner === currentPlayerName) {
-      // Player clicked on their own unit - select it using numeric ID
-      selectUnit(unitOnCell.id, clickedCellId);
-    } else if (selectedUnitId !== null) {
-      // Player has a unit selected - try to move it
-      moveUnit(selectedUnitId, selectedCellId!, clickedCellId);
-    } else {
-      // No unit selected and no unit on cell - try to place new unit
-      const cellOwner = currentTerritoryData[clickedCellId];
-      
-      if (cellOwner === currentPlayerName) {
-        placeEntity(clickedCellId);
-      } else if (cellOwner) {
-        showGameNotification(`This cell belongs to ${cellOwner}`, 'warning');
-      } else {
-        showGameNotification('You can only place units on your territory', 'warning');
-      }
-    }
-  }
-}
-
-function findUnitOnCell(cellId: number): any | null {
-  for (const [entityId, entity] of Object.entries(currentGameEntities)) {
-    if (entity.cellId === cellId && entity.type === 'unit') {
-      return entity; // Entity already has correct numeric id
-    }
-  }
-  return null;
-}
-
-function selectUnit(unitId: number, cellId: number): void {
-  selectedUnitId = unitId; // Keep as number
-  selectedCellId = cellId;
-
-  console.log(`NEW SELECTED UNIT ID: ${selectedUnitId}`)
-  
-  // Re-render to show selection highlight
-  renderGameState();
-}
-
-function deselectUnit(): void {
-  selectedUnitId = null;
-  selectedCellId = null;
-  renderGameState();
-}
-
-function moveUnit(unitId: number, fromCellId: number, toCellId: number): void {
-  console.log('Sending move action:', { unitId, fromCellId, toCellId });
-  
-  // Send move action to server with all numeric IDs
-  dispatchGameAction('move_unit', {
-    unitId,
-    fromCellId,
-    toCellId,
-  });
-  
-  // Deselect the unit
-  deselectUnit();
-  
-  console.log(`Sending move_unit action: unit ${unitId} from cell ${fromCellId} to cell ${toCellId}`);
-}
-
-function findCellAtPosition(x: number, y: number): number {
-  if (!meshData) return -1;
-
-  const nCells = meshData.cellCount;
-
-  for (let cellId = 0; cellId < nCells; cellId++) {
-    const start = meshData.cellOffsets[cellId];
-    const end = meshData.cellOffsets[cellId + 1];
-    
-    if (start >= end) continue;
-
-    // Check if point is inside this cell using ray casting algorithm
-    if (isPointInCell(x, y, cellId)) {
-      return cellId;
-    }
-  }
-
-  return -1;
-}
-
-function isPointInCell(x: number, y: number, cellId: number): boolean {
-  if (!meshData) return false;
-
-  const start = meshData.cellOffsets[cellId];
-  const end = meshData.cellOffsets[cellId + 1];
-  
-  if (start >= end) return false;
-
-  // Ray casting algorithm to check if point is inside polygon
-  let inside = false;
-  let j = end - 1;
-
-  for (let i = start; i < end; i++) {
-    const vi = meshData.cellVertexIndices[i];
-    const vj = meshData.cellVertexIndices[j];
-    
-    const xi = meshData.allVertices[vi * 2];
-    const yi = meshData.allVertices[vi * 2 + 1];
-    const xj = meshData.allVertices[vj * 2];
-    const yj = meshData.allVertices[vj * 2 + 1];
-
-    if (((yi > y) !== (yj > y)) && (x < (xj - xi) * (y - yi) / (yj - yi) + xi)) {
-      inside = !inside;
-    }
-    j = i;
-  }
-
-  return inside;
-}
-
-function placeEntity(cellId: number): void {
-  dispatchGameAction('place_entity', {
-    cellId,
-    entityType: 'unit',
-  });
-
-  console.log(`Sending place_entity action for cell ${cellId} by ${currentPlayerName}`);
-}
-
-function drawEntityMarker(cellId: number, owner: string, isSelected: boolean = false): void {
-  if (!meshData) return;
-
-  // Get the center of the cell for placing the entity marker
-  const centerX = meshData.cellTriangleCenters[cellId * 2];
-  const centerY = meshData.cellTriangleCenters[cellId * 2 + 1];
-
-  // Use different colors for different players
-  const playerColors: { [playerId: string]: string } = {
-    'player1': '#FF0000', // Red
-    'player2': '#0000FF', // Blue
-    'player3': '#00FF00', // Green
-    'player4': '#FFFF00', // Yellow
-    'player5': '#FF00FF', // Magenta
-    'player6': '#00FFFF', // Cyan
-  };
-  
-  const fillColor = playerColors[owner] || '#888888';
-  
-  // Draw selection ring if selected
-  if (isSelected) {
-    console.log(`UNIT ON ${cellId} selected`)
-    ctx.strokeStyle = '#FFFFFF';
-    ctx.lineWidth = 4;
-    ctx.beginPath();
-    ctx.arc(centerX, centerY, 12, 0, 2 * Math.PI);
-    ctx.stroke();
-  }
-  
-  ctx.fillStyle = fillColor;
-  ctx.strokeStyle = '#000000';
-  ctx.lineWidth = 2;
-  
-  ctx.beginPath();
-  ctx.arc(centerX, centerY, 8, 0, 2 * Math.PI);
-  ctx.fill();
-  ctx.stroke();
-  
-  // Add a small indicator for the unit type
-  ctx.fillStyle = '#FFFFFF';
-  ctx.font = '10px Arial';
-  ctx.textAlign = 'center';
-  ctx.textBaseline = 'middle';
-  ctx.fillText('⚔️', centerX, centerY);
-}
-
-// Event listener setup function to be called during initialization
-function setupGameplayEventListeners(): void {
-  // Add click handler to canvas for entity placement and movement
-  canvas.addEventListener('click', handleCellClick);
-  
-  // Add keyboard handler to deselect units
-  document.addEventListener('keydown', (event) => {
-    if (event.key === 'Escape' && selectedUnitId !== null) {
-      deselectUnit();
-      showGameNotification('Unit deselected', 'success');
-    }
-  });
-}
-
-// Handle errors
-function handleGameError(data: { error: string, gameId?: string }) {
-  console.error('Game error:', data);
-  
-  if (!data.gameId || data.gameId === currentGameId) {
-    showGameNotification(data.error, 'error');
-  }
-}
-
-// Manage game room membership
-function joinGameRoom(gameId: string, playerName: string, creator: boolean = false) {
-  currentGameId = gameId;
-  currentPlayerName = playerName;
-  isGameCreator = creator;
-  addToRoom(gameId, playerName, creator);
-  console.log(`Added ws to game room: ${gameId} as ${playerName} (creator: ${creator})`);
-}
-
-function leaveGameRoom() {
-  removeFromRoom(currentGameId, currentPlayerName);
-  if (currentGameId) {
-    console.log(`Removed ws from game room: ${currentGameId}`);
-  }
-  currentGameId = null;
-  currentPlayerName = null;
-  isGameCreator = false;
-}
-
-function dispatchGameAction(actionType: string, actionData: any) {
-  if (!currentGameId || !currentPlayerName) {
-    console.error('Missing game context - cannot send action');
-    return;
-  }
-  sendGameAction(actionType, {
-    gameId: currentGameId,
-    playerId: currentPlayerName,
-    ...actionData,
-  });
 }
 
 // Update game status display
@@ -905,185 +532,6 @@ function updateGameStatus(status: string) {
       statusElement.style.color = '#FFA500';
   }
 }
-
-/**
- * Processes game data received from WebSocket (complete Game object)
- */
-function processGameData(gameData: any): void {
-  try {
-    // Store game information from meta
-    currentGameId = gameData.meta.gameId;
-    
-    // Process territory data from game state
-    currentTerritoryData = gameData.state.cellOwnership || {};
-    
-    // Update turn state
-    isMyTurn = gameData.state.currentPlayer === currentPlayerName;
-    updateTurnIndicator(gameData.state.currentPlayer, gameData.state.turnNumber);
-    
-    // Update entities if provided
-    if (gameData.state.entities) {
-      currentGameEntities = gameData.state.entities;
-      console.log('Initial entities:', currentGameEntities);
-    }
-    
-    // Deserialize terrain data from game map
-    let terrainData: Uint8Array;
-    if (gameData.map.biomes.__typedArray) {
-      // Handle TypedArray descriptor from server
-      terrainData = deserializeTypedArrays(gameData.map.biomes);
-    } else {
-      // Handle direct Uint8Array
-      terrainData = new Uint8Array(gameData.map.biomes);
-    }
-    
-    currentGameTerrain = terrainData;
-    
-    console.log(`✅ Game data processed: ${terrainData.length} cells, ${Object.keys(currentTerritoryData).length} owned cells`);
-    console.log(`Current player: ${gameData.state.currentPlayer}, My turn: ${isMyTurn}`);
-    
-    // Update the current biomes with the game terrain
-    setCurrentCellBiomes(terrainData);
-    setCurrentCellCount(terrainData.length);
-
-    // Update the map size to match the game
-    setCurrentMapSize(gameData.meta.mapSize as MapSize);
-    
-    // Load the appropriate mesh if we don't have it
-    if (!meshData || meshData.cellCount !== terrainData.length) {
-      console.log(`Loading ${gameData.meta.mapSize} mesh for game...`);
-      loadOrGetMesh(gameData.meta.mapSize as MapSize, ctx).then(() => {
-        renderGameState();
-      });
-    } else {
-      // Render immediately if we have the right mesh
-      renderGameState();
-    }
-    
-  } catch (error: any) {
-    console.error('Failed to process game data:', error);
-    showError(`Failed to process game data: ${error.message}`);
-  }
-}
-
-/**
- * Renders the current game state with terrain and territories
- */
-function renderGameState(): void {
-  if (!meshData || !currentGameTerrain) {
-    console.warn('Cannot render game state: missing mesh data or terrain');
-    return;
-  }
-  
-  console.time('renderGameState');
-  
-  // Use the game terrain instead of generated terrain
-  const cellBiomes = currentGameTerrain;
-  
-  // Draw the base terrain
-  drawCells(
-    WIDTH,
-    HEIGHT,
-    ctx, 
-    meshData.allVertices, 
-    meshData.cellOffsets, 
-    meshData.cellVertexIndices, 
-    cellBiomes,
-    meshData.cellNeighbors,
-    biomeConfig.smoothColors
-  );
-  
-  // Overlay territory ownership
-  drawTerritoryOverlay();
-  
-  // Draw entities if we have entity data
-  if (currentGameEntities) {
-    drawEntities();
-  }
-  
-  console.timeEnd('renderGameState');
-}
-
-function drawEntities(): void {
-  if (!meshData) return;
-  
-  for (const [entityId, entity] of Object.entries(currentGameEntities)) {
-    if (entity.type === 'unit') {
-      // Compare numeric selectedUnitId with numeric entityId
-      const isSelected = selectedUnitId === parseInt(entityId);
-      console.log(`DRAWING UNIT ${entityId} (selected = ${isSelected})`)
-      drawEntityMarker(entity.cellId, entity.owner, isSelected);
-    }
-  }
-}
-
-/**
- * Draws territory ownership overlay on the map
- */
-function drawTerritoryOverlay(): void {
-  if (!meshData || Object.keys(currentTerritoryData).length === 0) {
-    return;
-  }
-  
-  // Territory colors for each player
-  const territoryColors: { [playerId: string]: string } = {
-    'player1': 'rgba(255, 0, 0, 0.3)',   // Red
-    'player2': 'rgba(0, 0, 255, 0.3)',   // Blue
-    'player3': 'rgba(0, 255, 0, 0.3)',   // Green
-    'player4': 'rgba(255, 255, 0, 0.3)', // Yellow
-    'player5': 'rgba(255, 0, 255, 0.3)', // Magenta
-    'player6': 'rgba(0, 255, 255, 0.3)', // Cyan
-  };
-  
-  // Draw territory ownership
-  for (const [cellIdStr, playerId] of Object.entries(currentTerritoryData)) {
-    const cellId = parseInt(cellIdStr);
-    const color = territoryColors[playerId] || 'rgba(128, 128, 128, 0.3)';
-    
-    // Get cell boundaries
-    const start = meshData.cellOffsets[cellId];
-    const end = meshData.cellOffsets[cellId + 1];
-    
-    if (start >= end) continue;
-    
-    // Draw territory overlay
-    ctx.fillStyle = color;
-    ctx.beginPath();
-    const v0 = meshData.cellVertexIndices[start];
-    ctx.moveTo(meshData.allVertices[v0 * 2], meshData.allVertices[v0 * 2 + 1]);
-    
-    for (let j = start + 1; j < end; j++) {
-      const vi = meshData.cellVertexIndices[j];
-      ctx.lineTo(meshData.allVertices[vi * 2], meshData.allVertices[vi * 2 + 1]);
-    }
-    
-    ctx.closePath();
-    ctx.fill();
-  }
-}
-
-
-// ===============================================================================================
-// ======================================== Setup Page ===========================================
-// ===============================================================================================
-
-
-let isMyTurn: boolean = false;
-
-let selectedUnitId: number | null = null;
-let selectedCellId: number | null = null;
-
-let currentGameEntities: { [entityId: number]: any } = {};
-
-let currentTerritoryData: { [cellId: number]: string } = {};
-let currentGameTerrain: Uint8Array | null = null;
-
-// Game state
-let currentGameId: string | null = null;
-let currentPlayerName: string | null = null;
-let isGameCreator: boolean = false;
-
-// Terrain configuration and canvas setup moved to separate modules
 
 // Start the application
 initializeApp().catch(error => {
